app/whisper-large_v3/requirements.txt
faster-whisper>=1.0.0
google-generativeai>=0.7.0
python-dotenv>=1.0.0
